Changed logic on startPage
The code did not work as it should on the startPage so we changed that. The idea from the beginning didn't work so we had to change so we only get the weather for one specific day.

We also changed so that the code have different functions instead of having all the things it does under one function.

Co-Authored-By: kroonfilip <81237654+kroonfilip@users.noreply.github.com>